implemented dynamic updater for news page
news page will now poll for new news every second while the popup is
open. every second it checks for a boolean indicating whether or not a
new item has been detected. If so, and the news page is rendering, then
it will update the news page and reset the boolean